// src/typings/Support.ts
export type Support = {
  分野?: string;
  サービス分類?: string;
  キーワード?: string;
  サービス名称?: string;
  対象者?: string;
  開始日?: string;
  終了日?: string;
  期間備考?: string;
  '無料/有料'?: string;
  導入支援有無?: string;
  企業等?: string;
  詳細?: string;
  発表日付?: string;
  URL?: string;
  情報源?: string;
};